// src/es/schemas/_types.query_dsl.ts
export const QueryDslBoolQuery = z.object({
  boost: z.lazy(() => float).describe('Floating point number used to decrease or increase the relevance scores of the query. Boost values are relative to the default value of 1.0. A boost value between 0 and 1.0 decreases the relevance score. A value greater than 1.0 increases the relevance score.').optional(),
  query_name: z.string().optional(),
  get filter (): z.ZodOptional<z.ZodUnion<readonly [typeof QueryDslQueryContainer, z.ZodArray<typeof QueryDslQueryContainer>]>> { return z.union([QueryDslQueryContainer, QueryDslQueryContainer.array()]).describe('The clause (query) must appear in matching documents. However, unlike `must`, the score of the query will be ignored.').optional() },
  minimum_should_match: z.lazy(() => MinimumShouldMatch).describe('Specifies the number or percentage of `should` clauses returned documents must match.').optional(),
  get must (): z.ZodOptional<z.ZodUnion<readonly [typeof QueryDslQueryContainer, z.ZodArray<typeof QueryDslQueryContainer>]>> { return z.union([QueryDslQueryContainer, QueryDslQueryContainer.array()]).describe('The clause (query) must appear in matching documents and will contribute to the score.').optional() },
  get must_not (): z.ZodOptional<z.ZodUnion<readonly [typeof QueryDslQueryContainer, z.ZodArray<typeof QueryDslQueryContainer>]>> { return z.union([QueryDslQueryContainer, QueryDslQueryContainer.array()]).describe('The clause (query) must not appear in the matching documents. Because scoring is ignored, a score of `0` is returned for all documents.').optional() },
  get should (): z.ZodOptional<z.ZodUnion<readonly [typeof QueryDslQueryContainer, z.ZodArray<typeof QueryDslQueryContainer>]>> { return z.union([QueryDslQueryContainer, QueryDslQueryContainer.array()]).describe('The clause (query) should appear in the matching document.').optional() }
}).meta({ id: 'QueryDslBoolQuery' })
export type QueryDslBoolQuery = z.infer<typeof QueryDslBoolQuery>

export interface QueryDslBoostingQueryShape {
  boost?: float | undefined
  query_name?: string | undefined
  negative_boost: double
  negative: QueryDslQueryContainerShape
  positive: QueryDslQueryContainerShape
}
export const QueryDslBoostingQuery = z.object({
  boost: z.lazy(() => float).describe('Floating point number used to decrease or increase the relevance scores of the query. Boost values are relative to the default value of 1.0. A boost value between 0 and 1.0 decreases the relevance score. A value greater than 1.0 increases the relevance score.').optional(),
  query_name: z.string().optional(),
  negative_boost: z.lazy(() => double).describe('Floating point number between 0 and 1.0 used to decrease the relevance scores of documents matching the `negative` query.'),
  get negative () { return QueryDslQueryContainer.describe('Query used to decrease the relevance score of matching documents.') },
  get positive () { return QueryDslQueryContainer.describe('Any returned documents must match this query.') }
}).meta({ id: 'QueryDslBoostingQuery' })
export type QueryDslBoostingQuery = z.infer<typeof QueryDslBoostingQuery>

export const QueryDslOperator = z.enum(['and', 'AND', 'or', 'OR']).meta({ id: 'QueryDslOperator' })
export type QueryDslOperator = z.infer<typeof QueryDslOperator>

export const QueryDslCommonTermsQuery = z.object({
  boost: z.lazy(() => float).describe('Floating point number used to decrease or increase the relevance scores of the query. Boost values are relative to the default value of 1.0. A boost value between 0 and 1.0 decreases the relevance score. A value greater than 1.0 increases the relevance score.').optional(),
  query_name: z.string().optional(),
  analyzer: z.string().optional(),
  cutoff_frequency: z.lazy(() => double).optional(),
  high_freq_operator: z.lazy(() => QueryDslOperator).optional(),
  low_freq_operator: z.lazy(() => QueryDslOperator).optional(),
  minimum_should_match: z.lazy(() => MinimumShouldMatch).optional(),
  query: z.string()
}).meta({ id: 'QueryDslCommonTermsQuery' })
export type QueryDslCommonTermsQuery = z.infer<typeof QueryDslCommonTermsQuery>
// ...
const QueryDslQueryContainerExclusiveProps = z.union([z.object({ bool: z.lazy(() => QueryDslBoolQuery) }), z.object({ boosting: z.lazy(() => QueryDslBoostingQuery) }), z.object({ common: z.record(z.lazy(() => Field), z.union([QueryDslCommonTermsQuery, z.string()])) }), z.object({ combined_fields: QueryDslCombinedFieldsQuery }), z.object({ constant_score: z.lazy(() => QueryDslConstantScoreQuery) }), z.object({ dis_max: z.lazy(() => QueryDslDisMaxQuery) }), z.object({ distance_feature: QueryDslDistanceFeatureQuery }), z.object({ exists: z.lazy(() => QueryDslExistsQuery) }), z.object({ function_score: z.union([z.lazy(() => QueryDslFunctionScoreQuery), z.array(z.lazy(() => QueryDslFunctionScoreContainer))]) }), z.object({ fuzzy: z.record(z.lazy(() => Field), z.union([QueryDslFuzzyQuery, z.union([z.string(), z.lazy(() => double), z.boolean()])])) }), z.object({ geo_bounding_box: QueryDslGeoBoundingBoxQuery }), z.object({ geo_distance: QueryDslGeoDistanceQuery }), z.object({ geo_grid: z.record(z.lazy(() => Field), QueryDslGeoGridQuery) }), z.object({ geo_polygon: QueryDslGeoPolygonQuery }), z.object({ geo_shape: QueryDslGeoShapeQuery }), z.object({ has_child: z.lazy(() => QueryDslHasChildQuery) }), z.object({ has_parent: z.lazy(() => QueryDslHasParentQuery) }), z.object({ ids: z.lazy(() => QueryDslIdsQuery) }), z.object({ intervals: z.record(z.lazy(() => Field), z.lazy(() => QueryDslIntervalsQuery)) }), z.object({ knn: z.lazy(() => KnnQuery) }), z.object({ match: z.record(z.lazy(() => Field), z.union([z.lazy(() => QueryDslMatchQuery), z.union([z.string(), z.lazy(() => float), z.boolean()])])) }), z.object({ match_all: z.lazy(() => QueryDslMatchAllQuery) }), z.object({ match_bool_prefix: z.record(z.lazy(() => Field), z.union([QueryDslMatchBoolPrefixQuery, z.string()])) }), z.object({ match_none: QueryDslMatchNoneQuery }), z.object({ match_phrase: z.record(z.lazy(() => Field), z.union([QueryDslMatchPhraseQuery, z.string()])) }), z.object({ match_phrase_prefix: z.record(z.lazy(() => Field), z.union([QueryDslMatchPhrasePrefixQuery, z.string()])) }), z.object({ more_like_this: QueryDslMoreLikeThisQuery }), z.object({ multi_match: QueryDslMultiMatchQuery }), z.object({ nested: z.lazy(() => QueryDslNestedQuery) }), z.object({ parent_id: QueryDslParentIdQuery }), z.object({ percolate: QueryDslPercolateQuery }), z.object({ pinned: z.lazy(() => QueryDslPinnedQuery) }), z.object({ prefix: z.record(z.lazy(() => Field), z.union([z.lazy(() => QueryDslPrefixQuery), z.string()])) }), z.object({ query_string: QueryDslQueryStringQuery }), z.object({ range: z.record(z.lazy(() => Field), z.lazy(() => QueryDslRangeQuery)) }), z.object({ rank_feature: QueryDslRankFeatureQuery }), z.object({ regexp: z.record(z.lazy(() => Field), z.union([QueryDslRegexpQuery, z.string()])) }), z.object({ rule: z.lazy(() => QueryDslRuleQuery) }), z.object({ script: z.lazy(() => QueryDslScriptQuery) }), z.object({ script_score: z.lazy(() => QueryDslScriptScoreQuery) }), z.object({ semantic: QueryDslSemanticQuery }), z.object({ shape: QueryDslShapeQuery }), z.object({ simple_query_string: z.lazy(() => QueryDslSimpleQueryStringQuery) }), z.object({ span_containing: z.lazy(() => QueryDslSpanContainingQuery) }), z.object({ span_field_masking: z.lazy(() => QueryDslSpanFieldMaskingQuery) }), z.object({ span_first: z.lazy(() => QueryDslSpanFirstQuery) }), z.object({ span_multi: z.lazy(() => QueryDslSpanMultiTermQuery) }), z.object({ span_near: z.lazy(() => QueryDslSpanNearQuery) }), z.object({ span_not: z.lazy(() => QueryDslSpanNotQuery) }), z.object({ span_or: z.lazy(() => QueryDslSpanOrQuery) }), z.object({ span_term: z.record(z.lazy(() => Field), z.union([QueryDslSpanTermQuery, z.lazy(() => FieldValue)])) }), z.object({ span_within: z.lazy(() => QueryDslSpanWithinQuery) }), z.object({ sparse_vector: QueryDslSparseVectorQuery }), z.object({ term: z.record(z.lazy(() => Field), z.union([z.lazy(() => QueryDslTermQuery), z.lazy(() => FieldValue)])) }), z.object({ terms: z.lazy(() => QueryDslTermsQuery) }), z.object({ terms_set: z.record(z.lazy(() => Field), z.lazy(() => QueryDslTermsSetQuery)) }), z.object({ text_expansion: z.record(z.lazy(() => Field), QueryDslTextExpansionQuery) }), z.object({ weighted_tokens: z.record(z.lazy(() => Field), QueryDslWeightedTokensQuery) }), z.object({ wildcard: z.record(z.lazy(() => Field), z.union([z.lazy(() => QueryDslWildcardQuery), z.string()])) }), z.object({ wrapper: QueryDslWrapperQuery }), z.object({ type: QueryDslTypeQuery })])

export interface QueryDslQueryContainerShape {
  bool?: QueryDslBoolQuery | undefined
  boosting?: QueryDslBoostingQuery | undefined
  common?: Record<Field, QueryDslCommonTermsQuery | string> | undefined
  combined_fields?: QueryDslCombinedFieldsQuery | undefined
  constant_score?: QueryDslConstantScoreQuery | undefined
  dis_max?: QueryDslDisMaxQuery | undefined
  distance_feature?: QueryDslDistanceFeatureQuery | undefined
  exists?: QueryDslExistsQuery | undefined
  function_score?: QueryDslFunctionScoreQuery | QueryDslFunctionScoreContainer[] | undefined
  fuzzy?: Record<Field, QueryDslFuzzyQuery | string | double | boolean> | undefined
  geo_bounding_box?: QueryDslGeoBoundingBoxQuery | undefined
  geo_distance?: QueryDslGeoDistanceQuery | undefined
  geo_grid?: Record<Field, QueryDslGeoGridQuery> | undefined
  geo_polygon?: QueryDslGeoPolygonQuery | undefined
  geo_shape?: QueryDslGeoShapeQuery | undefined
  has_child?: QueryDslHasChildQuery | undefined
  has_parent?: QueryDslHasParentQuery | undefined
  ids?: QueryDslIdsQuery | undefined
  intervals?: Record<Field, QueryDslIntervalsQuery> | undefined
  knn?: KnnQuery | undefined
  match?: Record<Field, QueryDslMatchQuery | string | float | boolean> | undefined
  match_all?: QueryDslMatchAllQuery | undefined
  match_bool_prefix?: Record<Field, QueryDslMatchBoolPrefixQuery | string> | undefined
  match_none?: QueryDslMatchNoneQuery | undefined
  match_phrase?: Record<Field, QueryDslMatchPhraseQuery | string> | undefined
  match_phrase_prefix?: Record<Field, QueryDslMatchPhrasePrefixQuery | string> | undefined
  more_like_this?: QueryDslMoreLikeThisQuery | undefined
  multi_match?: QueryDslMultiMatchQuery | undefined
  nested?: QueryDslNestedQuery | undefined
  parent_id?: QueryDslParentIdQuery | undefined
  percolate?: QueryDslPercolateQuery | undefined
  pinned?: QueryDslPinnedQuery | undefined
  prefix?: Record<Field, QueryDslPrefixQuery | string> | undefined
  query_string?: QueryDslQueryStringQuery | undefined
  range?: Record<Field, QueryDslRangeQuery> | undefined
  rank_feature?: QueryDslRankFeatureQuery | undefined
  regexp?: Record<Field, QueryDslRegexpQuery | string> | undefined
  rule?: QueryDslRuleQuery | undefined
  script?: QueryDslScriptQuery | undefined
  script_score?: QueryDslScriptScoreQuery | undefined
  semantic?: QueryDslSemanticQuery | undefined
  shape?: QueryDslShapeQuery | undefined
  simple_query_string?: QueryDslSimpleQueryStringQuery | undefined
  span_containing?: QueryDslSpanContainingQuery | undefined
  span_field_masking?: QueryDslSpanFieldMaskingQuery | undefined
  span_first?: QueryDslSpanFirstQuery | undefined
  span_multi?: QueryDslSpanMultiTermQuery | undefined
  span_near?: QueryDslSpanNearQuery | undefined
  span_not?: QueryDslSpanNotQuery | undefined
  span_or?: QueryDslSpanOrQuery | undefined
  span_term?: Record<Field, QueryDslSpanTermQuery | FieldValue> | undefined
  span_within?: QueryDslSpanWithinQuery | undefined
  sparse_vector?: QueryDslSparseVectorQuery | undefined
  term?: Record<Field, QueryDslTermQuery | FieldValue> | undefined
  terms?: QueryDslTermsQuery | undefined
  terms_set?: Record<Field, QueryDslTermsSetQuery> | undefined
  text_expansion?: Record<Field, QueryDslTextExpansionQuery> | undefined
  weighted_tokens?: Record<Field, QueryDslWeightedTokensQuery> | undefined
  wildcard?: Record<Field, QueryDslWildcardQuery | string> | undefined
  wrapper?: QueryDslWrapperQuery | undefined
  type?: QueryDslTypeQuery | undefined
}
/** An Elasticsearch Query DSL (Domain Specific Language) object that defines a query. */
export const QueryDslQueryContainer: z.ZodType<QueryDslQueryContainerShape> = QueryDslQueryContainerExclusiveProps.meta({ id: 'QueryDslQueryContainer' })
export type QueryDslQueryContainer = z.infer<typeof QueryDslQueryContainer>